+++ methods/to_text.ts
@@ -1,0 +1,8 @@
+import type { ElementReference } from "../types.ts";
+
+/**
+ * Return a valid SVG text representation of the given element.
+ */
+export function toText(_element: ElementReference): string {
+  throw new Error("not implemented yet");
+}
